// src/modules/rooms/room.controller.d.ts
export declare const roomController: {
    createRoom: (req: import("express").Request, res: import("express").Response, next: import("express").NextFunction) => void;
    joinRoom: (req: import("express").Request, res: import("express").Response, next: import("express").NextFunction) => void;
    leaveRoom: (req: import("express").Request, res: import("express").Response, next: import("express").NextFunction) => void;
    getRoom: (req: import("express").Request, res: import("express").Response, next: import("express").NextFunction) => void;
    getRiders: (req: import("express").Request, res: import("express").Response, next: import("express").NextFunction) => void;
    getMyRooms: (req: import("express").Request, res: import("express").Response, next: import("express").NextFunction) => void;
    getRideHistory: (req: import("express").Request, res: import("express").Response, next: import("express").NextFunction) => void;
    closeRoom: (req: import("express").Request, res: import("express").Response, next: import("express").NextFunction) => void;
};
//# sourceMappingURL=room.controller.d.ts.map